requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
biopython>=1.79
scikit-learn>=0.24.0 